Retry PR (#2)
* Implement the retry of connections

When using multiple proxy upstreams, when one connection fails becuase the proxy is not available, then try the next until check all available.

With this new behaviour, you never receive an error if almost one proxy is reachable. Useful in environments when multiple non-reliable proxies are used.

* fix potential edge case

Add a set to overcome the case when seed is incorrect